requirements.txt
# Google Cloud and ADK dependencies
google-adk>=0.1.0
google-cloud-aiplatform>=1.38.0
google-cloud-bigquery>=3.11.4
google-cloud-firestore>=2.11.1
google-generativeai>=0.3.0
vertexai>=1.38.0

# Maps and Location APIs
googlemaps>=4.10.0

# Weather API
requests>=2.31.0

# Payment processing
stripe>=7.0.0

# Data validation and modeling
pydantic>=2.4.0

# Type hints and utilities
typing-extensions>=4.8.0

# Date and time handling
python-dateutil>=2.8.2

# Math and optimization
numpy>=1.24.0
scipy>=1.11.0

# Async support
asyncio
aiohttp>=3.8.0

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.1.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development dependencies
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Web server for local testing
uvicorn>=0.23.0
fastapi>=0.103.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Environment and configuration
python-dotenv>=1.0.0
pyyaml>=6.0

# Logging and monitoring
structlog>=23.1.0